Move files in unittest (#13)
* Merge devel to main to solve pyright error (#9)

* init devel

* Fix pyright error (#8)

Fix pyright error caused by the version update of pyright.

---------


---------


* Devel to main (#11)

* init devel

* Fix pyright error (#8)

Fix pyright error caused by the version update of pyright.

---------


---------


* Update README.md

* Update preprunvasp.py

* Create pub-pypi.yml (#12)

* move file

---------
